精简缓存逻辑
Fix https://github.com/HenryQW/RSSHub/issues/6

Fix https://github.com/HenryQW/RSSHub/issues/7
