+++ middleware/lru-cache.js
@@ -1,9 +1,7 @@
-// baed on https://github.com/coderhaoxin/koa-redis-cache
+// based on https://github.com/coderhaoxin/koa-redis-cache
 
-const pathToRegExp = require('path-to-regexp');
-const readall = require('readall');
-const crypto = require('crypto');
 const lru = require('lru-cache');
+const common = require('./cache-common');
 
 module.exports = function(options = {}) {
     const {
@@ -29,31 +27,10 @@ module.exports = function(options = {}) {
     return async function cache(ctx, next) {
         const { url, path } = ctx.request;
         const resolvedPrefix = typeof prefix === 'function' ? prefix.call(ctx, ctx) : prefix;
-        const key = resolvedPrefix + md5(ignoreQuery ? path : url);
+        const key = resolvedPrefix + common.md5(ignoreQuery ? path : url);
         const tkey = key + ':type';
-        let match = false;
-        let routeExpire = false;
 
-        for (let i = 0; i < routes.length; i++) {
-            let route = routes[i];
-
-            if (typeof routes[i] === 'object') {
-                route = routes[i].path;
-                routeExpire = routes[i].expire;
-            }
-
-            if (paired(route, path)) {
-                match = true;
-                break;
-            }
-        }
-
-        for (let j = 0; j < exclude.length; j++) {
-            if (paired(exclude[j], path)) {
-                match = false;
-                break;
-            }
-        }
+        let { match, routeExpire } = common.validityCheck(routes, exclude, path);
 
         if (!match || (passParam && ctx.request.query[passParam])) {
             return await next();
@@ -134,32 +111,3 @@ module.exports = function(options = {}) {
         }
     }
 };
-
-function paired(route, path) {
-    const options = {
-        sensitive: true,
-        strict: true,
-    };
-
-    return pathToRegExp(route, [], options).exec(path);
-}
-
-// eslint-disable-next-line no-unused-vars
-function read(stream) {
-    return new Promise((resolve, reject) => {
-        readall(stream, (err, data) => {
-            if (err) {
-                reject(err);
-            } else {
-                resolve(data);
-            }
-        });
-    });
-}
-
-function md5(str) {
-    return crypto
-        .createHash('md5')
-        .update(str)
-        .digest('hex');
-}
